separate src into models.py and fitting_functions.py; update tutorial.ipynb to relfect changes
