+++ src/growthcurves/__init__.py
@@ -5,7 +5,8 @@ from importlib import metadata
 __version__ = metadata.version("growthcurves")
 
 from . import models
+from . import fitting_functions
 
 # The __all__ variable is a list of variables which are imported
 # when a user does "from example import *"
-__all__ = ["models"]
+__all__ = ["models", "fitting_functions"]
